% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SN_proteome}
\alias{SN_proteome}
\alias{SN_proteome_50}
\alias{SN_proteome_no_batch_effect}
\alias{SN_proteome_no_bias}
\alias{SN_proteome_residuals}
\title{MSnSet objects with preprocessed bottom-up proteomics data 
       for the Lewy bodies study}
\format{an \code{MSnSet} instance}
\source{
PNNL's data analysis pipeline
}
\usage{
data(SN_proteome)
data(SN_proteome_50)
data(SN_proteome_no_batch_effect)
data(SN_proteome_no_bias)
data(SN_proteome_residuals)
}
\value{
m an \code{MSnSet} instance
}
\description{
Bottom-up LC-MS/MS data collected on the Q Exactive
             instrument. Mass spec data was analyzed by MaxQuant.
}
\details{
There are four MSnSet objects. The list of objects is 
ordered according to the pre-processing steps.
\itemize{
  \item `SN_proteome`: LFQ intensities extracted from MaxQuant output.
                     Preprocessing involves log2 transform and 
                     zero-centering for proteins.

  \item `SN_proteome_50`: Only those proteins that 
  present in at least 50% of samples are retained.
  Then normalized using \code{\link[stats]{medpolish}} approach.
  
  \item `SN_proteome_no_batch_effect`: All the previous steps, 
  then technical batch effect correspoding to plate columns 
  is modeled and removed. 
  Then normalized using \code{\link[stats]{medpolish}} approach.
  
  \item `SN_proteome_no_bias`: Same as `SN_proteome_no_batch_effect`,
  but additionally biases correspoding to subject matching 
  (age, gender, etc) are also removed.
  Then normalized using \code{\link[stats]{medpolish}} approach.
  
  \item `SN_proteome_residuals`: Same as `SN_proteome_no_bias`,
  but additionally group effects (case, control.1, control.2) are
  also removed.
  Then normalized using \code{\link[stats]{medpolish}} approach.
}
The details on data extraction and preprocessing can be found 
in the package source "scripts" directory.
}
\examples{
\dontrun{
data(SN_proteome)
head(pData(m))
head(fData(m))
head(exprs(m))
}
}
\author{
Vlad Petyuk, 2016-07-06
}

